+++ tests/test_file_data.py
@@ -4853,7 +4853,7 @@ class ModelImputationWithExtrapolationAndInterpolationData:
 
 
 @dataclass
-class ModelExtrapolationNew:
+class ModelExtrapolation:
     extrapolation_rows = [
         ("1-001", date(2023, 1, 1), 1672531200, 15.0, 15.0),
         ("1-001", date(2023, 2, 1), 1675209600, None, 15.1),
